// React/app/GerAdv_TS/AlertasEnviados/Interfaces/interface.AlertasEnviados.ts
// 100% auto-generated by Source Genesys WebApi Front & Back Creator
                     // copyright © 2000-2025 Menphis - Sistemas Inteligentes
                     // This file is part of the Source Genesys project                     
                    'use client';

export interface IAlertasEnviados {
// 202501251
    id: number;
 
	operador: number,
	alerta: number,
	dataalertado: string,
	visualizado: boolean,
}

export interface IAlertasEnviadosFormProps {
  id: number;
  onClose: () => void;
  onSuccess: (registro?: any) => void;
  onError: () => void;
}

export interface IAlertasEnviadosIncProps {
    id: number;
    onClose: () => void;
    onSuccess: (registro?: any) => void;
    onError: () => void;
}